+++ .idea/.gitignore
@@ -1,0 +1,3 @@
+# 默认忽略的文件
+/shelf/
+/workspace.xml
