--- CsvTransfer/1/1-backend/src/main/resources/application-default.properties
@@ -1,1 +1,0 @@
-logging.level.web=debug
