--- CsvTransfer/1/1-backend/src/test/java/pk/AppTest.java
@@ -1,12 +1,0 @@
-package pk;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AppTest {
-
-    @Test
-    void contextLoads() {
-    }
-}
