--- CsvTransfer/1/1-backend/src/test/resources/junit-platform.properties
@@ -1,1 +1,0 @@
-junit.jupiter.testinstance.lifecycle.default=per_class
