--- CsvTransfer/1/1-frontend/pom.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>1</artifactId>
-        <groupId>1</groupId>
-        <version>1</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>1-frontend</artifactId>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>net.unit8.maven.plugins</groupId>
-                <artifactId>elm-maven-plugin</artifactId>
-                <version>0.1.0</version>
-                <configuration>
-                    <installDirectory>target</installDirectory>
-                </configuration>
-                <executions>
-                    <execution>
-                        <id>install-elm</id>
-                        <goals>
-                            <goal>install</goal>
-                        </goals>
-                        <configuration>
-                            <elmVersion>0.19.0</elmVersion>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>elm-make</id>
-                        <goals>
-                            <goal>run</goal>
-                        </goals>
-                        <phase>compile</phase>
-                        <configuration>
-                            <arguments>make src/Main.elm --output=${project.build.directory}/classes/META-INF/resources/index.html</arguments>
-                            <workingDirectory>${basedir}</workingDirectory>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-</project>
