--- CsvTransfer/1/manifest.yml
@@ -1,7 +1,0 @@
-applications:
-- name: 1
-  path: 1-backend/target/1-backend-1.0-SNAPSHOT.jar
-  memory: 256m
-  env:
-    JAVA_OPTS: '-XX:ReservedCodeCacheSize=32M -XX:MaxDirectMemorySize=32M'
-    JBP_CONFIG_OPEN_JDK_JRE: '[memory_calculator: {stack_threads: 30}]'
