--- CsvTransfer/1/pom.xml
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <packaging>pom</packaging>
-  <parent>
-    <groupId>org.springframework.boot</groupId>
-    <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.2.0.M4</version>
-    <!-- lookup parent from repository -->
-  </parent>
-  <groupId>1</groupId>
-  <artifactId>1</artifactId>
-  <version>1</version>
-  <name>1</name>
-  <description>Blank multi project for Spring Boot + Elm</description>
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>1</groupId>
-        <artifactId>1-frontend</artifactId>
-        <version>1</version>
-      </dependency>
-      <dependency>
-        <groupId>1</groupId>
-        <artifactId>1-backend</artifactId>
-        <version>1</version>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-  <repositories>
-    <repository>
-      <id>spring-snapshots</id>
-      <name>Spring Snapshots</name>
-      <url>https://repo.spring.io/snapshot</url>
-      <snapshots/>
-    </repository>
-    <repository>
-      <id>spring-milestones</id>
-      <name>Spring Milestones</name>
-      <url>https://repo.spring.io/milestone</url>
-    </repository>
-  </repositories>
-  <pluginRepositories>
-    <pluginRepository>
-      <id>spring-snapshots</id>
-      <name>Spring Snapshots</name>
-      <url>https://repo.spring.io/snapshot</url>
-      <snapshots/>
-    </pluginRepository>
-    <pluginRepository>
-      <id>spring-milestones</id>
-      <name>Spring Milestones</name>
-      <url>https://repo.spring.io/milestone</url>
-    </pluginRepository>
-  </pluginRepositories>
-  <modules>
-    <module>1-frontend</module>
-    <module>1-backend</module>
-  </modules>
-</project>
