--- DeepSeekTest/src/test/java/com/example/deepseek/AppTest.java
@@ -1,20 +1,0 @@
-package com.example.deepseek;
-
-import static org.junit.Assert.assertTrue;
-
-import org.junit.Test;
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
