+++ src/main/java/utils/DeepSeek.java
@@ -1,0 +1,76 @@
+package utils;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ArrayNode;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import okhttp3.*;
+
+import java.io.IOException;
+import java.util.concurrent.TimeUnit;
+import java.util.logging.Logger;
+
+public class DeepSeek {
+    private static final Logger logger = Logger.getLogger(DeepSeek.class.getName());
+    private static final String API_KEY = ConfigUtil.getApiKey();
+    private static final String API_URL = "https://api.deepseek.com/chat/completions";
+
+    /**
+     * 和DeepSeek进行对话
+     *
+     * @param userInput 用户的输入（字符串）
+     * @return DeepSeek的返回内容
+     */
+    public static String communicate(String userInput) {
+        OkHttpClient client = new OkHttpClient().newBuilder()
+                .connectTimeout(30, TimeUnit.SECONDS)
+                .writeTimeout(60, TimeUnit.SECONDS)
+                .readTimeout(30, TimeUnit.SECONDS)
+                .build();
+        MediaType mediaType = MediaType.parse("application/json");
+        // 1. 构建 JSON 字符串（动态插入 userInput）
+        String jsonBody = "{\n" +
+                "  \"messages\": [\n" +
+                "    {\n" +
+                "      \"content\": \"You are a helpful assistant\",\n" +
+                "      \"role\": \"system\"\n" +
+                "    },\n" +
+                "    {\n" +
+                "      \"content\": \"" + userInput + "\",\n" +
+                "      \"role\": \"user\"\n" +
+                "    }\n" +
+                "  ],\n" +
+                "  \"model\": \"deepseek-chat\",\n" +
+                "  \"max_tokens\": 2048,\n" +
+                "  \"temperature\": 1,\n" +
+                "  \"tool_choice\": \"none\",\n" +
+                "  \"stream\": false\n" +
+                "}";
+
+        // 2. 发送请求
+        RequestBody body = RequestBody.create(jsonBody, mediaType);
+        Request request = new Request.Builder()
+                .url(API_URL)
+                .method("POST", body)
+                .addHeader("Content-Type", "application/json")
+                .addHeader("Accept", "application/json")
+                .addHeader("Authorization", "Bearer " + API_KEY)
+                .build();
+
+        try (Response response = client.newCall(request).execute()) {
+            if (!response.isSuccessful()) {
+                System.err.println("请求失败，状态码：" + response.code());
+                return "请求失败：" + response.body().string();
+            } else {
+                ObjectMapper mapper = new ObjectMapper();
+                JsonNode root = mapper.readTree(response.body().string());
+                JsonNode contentNode = root.path("choices").get(0).path("message").path("content");
+
+                return contentNode.asText();
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+            return "调用失败：" + e.getMessage();
+        }
+    }
+}
